+++ kernel/core/kentry.c
@@ -35,6 +35,9 @@
 #include <core/gdt.h>
 #include <core/syscall.h>
 #include <core/mm.h>
+#include <core/elf.h>
+#include <core/lock.h>
+#include <core/panic.h>
 
 #include <lib/kmemcpy.h>
 
@@ -73,6 +76,9 @@ extern uint64_t init_stack_paddr;
 extern uint64_t* init_stacks_paddr;
 extern uint64_t* init_stacks_vaddr;
 
+static uint8_t prepare_userland_lock;
+static uint8_t init_done;
+
 static inline void write_syscall_msr(void) {
 	msr_write(MSR_STAR, ((GDT_USER_CS - 0x10) << 48) | (GDT_KERNEL_CS << 32));
 	msr_write(MSR_LSTAR, (uint64_t)syscall_entry);
@@ -96,6 +102,8 @@ void kentry(void) {
 	scheduler_init();
 
 	mm_transaction_init();
+	init_done = 0;
+	lock_init(&prepare_userland_lock);
 
 	write_syscall_msr();
 	cpu_set_cr4(CR4_FSGSBASE);
@@ -181,4 +189,32 @@ void kapentry(uint64_t arb_id) {
 	logging_log_info("AP init complete");
 
 	process_kill_current();
+}
+
+void prepare_userland(void* cntx) {
+	(void)cntx;
+
+	lock_acquire(&prepare_userland_lock);
+	if (init_done) {
+		logging_log_error("Multiple calls to prepare userland");
+		panic(PANIC_STATE);
+	}
+
+	init_done = 1;
+	lock_release(&prepare_userland_lock);
+
+	struct fs_handle_t* shell = fs_open("/bin/shell");
+	if (!shell) {
+		logging_log_error("Failed to open shell file");
+	}
+
+	else {
+		struct pcb_t* shell_pcb = elf_load(shell, process_assign_pid(), "/bin/shell ModulOS", "USER=root PWD=/");
+		if (!shell_pcb) {
+			logging_log_error("Failed to load shell file");
+		}
+		fs_close(shell);
+
+		scheduler_schedule(shell_pcb);
+	}
 }
